move constants.h
